Support launchpad midi mapping.

+++ src/main/java/zone/kaz/alight_midi/device/midi/MidiData.java
@@ -19,7 +19,7 @@ public class MidiData {
         return type;
     }
 
-    public void setType(byte type) {
+    public void setType(int type) {
         this.type = type;
     }
